[NO-JIRA] Fix imports
