+++ helpers/constants.js
@@ -1,3 +1,11 @@
+const GENERIC_ERROR_MESSAGE = {
+  type: 'error',
+  message: 'Something went wrong. Please try again later.',
+};
+const COMMUNICATION_ERROR_MESSAGE = {
+  type: 'error',
+  message: 'Our servers are poorly. Please try again later.',
+};
 const PROJECT_NAME = 'GEORGEGILLAMS';
 const DECIMAL_REGEX = /^[0-9\.]*$/gi;
 const INT_REGEX = /^[0-9]*$/gi;
@@ -92,6 +100,7 @@ export {
   AWS,
   CARD_NUMBER_REGEX,
   CHECK_FOR_NEW_CONTENT_INTERVAL,
+  COMMUNICATION_ERROR_MESSAGE,
   COMPONENT_RELOAD_INTERVAL,
   CONFERENCE_DAY_1_END,
   CONFERENCE_DAY_1_START,
@@ -108,6 +117,7 @@ export {
   EMAIL_SENDER_EMAIL,
   EMAIL_VERIFICATION_ENABLED,
   EXPIRY_REGEX,
+  GENERIC_ERROR_MESSAGE,
   GG_EMAIL,
   ID_REGEX,
   INT_REGEX,
@@ -136,8 +146,8 @@ export {
   TICKET_COST_R_TWO_DAY,
   TICKET_RESERVATION_LENGTH,
   TICKET_SALE_END,
-  USERNAME_REGEX,
   USERNAMES_ENABLED,
+  USERNAME_REGEX,
 };
 export default {
   ANYTHING_REGEX,
@@ -146,6 +156,7 @@ export default {
   AWS,
   CARD_NUMBER_REGEX,
   CHECK_FOR_NEW_CONTENT_INTERVAL,
+  COMMUNICATION_ERROR_MESSAGE,
   COMPONENT_RELOAD_INTERVAL,
   CONFERENCE_DAY_1_END,
   CONFERENCE_DAY_1_START,
@@ -162,6 +173,7 @@ export default {
   EMAIL_SENDER_EMAIL,
   EMAIL_VERIFICATION_ENABLED,
   EXPIRY_REGEX,
+  GENERIC_ERROR_MESSAGE,
   GG_EMAIL,
   ID_REGEX,
   INT_REGEX,
@@ -190,6 +202,6 @@ export default {
   TICKET_COST_R_TWO_DAY,
   TICKET_RESERVATION_LENGTH,
   TICKET_SALE_END,
-  USERNAME_REGEX,
   USERNAMES_ENABLED,
+  USERNAME_REGEX,
 };
