+++ server/api/actions/auth/requestVerificationEmail.js
@@ -3,7 +3,7 @@ import authAllowedAttributes from './private/authAllowedAttributes';
 import authentication from 'utils/authentication';
 import reqSecure from 'utils/reqSecure';
 import { sendEmailVerificationEmail } from 'utils/emailHelpers';
-import { UNAUTHORISED_WRITE } from 'helpers/constants';
+import { UNAUTHORISED_WRITE } from 'utils/errorConstants';
 
 export default function requestVerificationEmail(req) {
   reqSecure(req, authAllowedAttributes);
