+++ server/api/actions/blogs/create.js
@@ -2,7 +2,7 @@ import { datumCreate } from '../datum';
 
 import lockPromise from 'utils/lock';
 import authentication from 'utils/authentication';
-import { UNAUTHORISED_WRITE } from 'helpers/constants';
+import { UNAUTHORISED_WRITE } from 'utils/errorConstants';
 
 export default function create(req) {
   return lockPromise('blogs', () =>
