+++ server/api/actions/blogs/delete.js
@@ -1,7 +1,7 @@
 import { datumRemove } from '../datum';
 
 import authentication from 'utils/authentication';
-import { UNAUTHORISED_WRITE } from 'helpers/constants';
+import { UNAUTHORISED_WRITE } from 'utils/errorConstants';
 
 export default function remove(req) {
   return authentication(req).then(user => {
