+++ server/api/actions/blogs/update.js
@@ -2,7 +2,7 @@ import { datumUpdate } from '../datum';
 
 import lockPromise from 'utils/lock';
 import authentication from 'utils/authentication';
-import { UNAUTHORISED_WRITE } from 'helpers/constants';
+import { UNAUTHORISED_WRITE } from 'utils/errorConstants';
 
 export default function update(req) {
   return lockPromise('blogs', () =>
