+++ server/api/actions/dataManagement/deleteEntity.js
@@ -1,12 +1,7 @@
 import { datumLoad } from '../datum';
 
-import {
-  STRING_REGEX,
-  ID_REGEX,
-  RESOURCE_NOT_FOUND,
-  PROJECT_NAME,
-  UNAUTHORISED_WRITE,
-} from 'helpers/constants';
+import { STRING_REGEX, ID_REGEX, PROJECT_NAME } from 'helpers/constants';
+import { RESOURCE_NOT_FOUND, UNAUTHORISED_WRITE } from 'utils/errorConstants';
 import { AuthError } from 'utils/errors';
 import redis from 'utils/redis';
 import { find } from 'utils/find';
