+++ server/api/actions/dataManagement/deleteSet.js
@@ -1,8 +1,5 @@
-import {
-  STRING_REGEX,
-  PROJECT_NAME,
-  UNAUTHORISED_WRITE,
-} from 'helpers/constants';
+import { STRING_REGEX, PROJECT_NAME } from 'helpers/constants';
+import { UNAUTHORISED_WRITE } from 'utils/errorConstants';
 import redis from 'utils/redis';
 import { InvalidInputError } from 'utils/errors';
 import authentication from 'utils/authentication';
