+++ server/api/actions/datum/datumRemove.js
@@ -1,14 +1,14 @@
-import { RESOURCE_NOT_FOUND } from 'helpers/constants';
-import { find } from 'utils/find';
-
 import datumLoad from './datumLoad';
 import datumUpdate from './datumUpdate';
+
+import { RESOURCE_NOT_FOUND } from 'utils/errorConstants';
+import { find } from 'utils/find';
 
 export default function datumRemove(settings, req) {
   return new Promise((resolve, reject) => {
     datumLoad(settings, req).then(
       data => {
-        const { existingValue, existingValueIndex } = find(data, req.body.id);
+        const { existingValue } = find(data, req.body.id);
 
         if (existingValue) {
           const value = JSON.parse(JSON.stringify(existingValue));
