+++ server/api/actions/notifications/delete.js
@@ -3,7 +3,7 @@ import { datumRemove } from '../datum';
 import notificationsAllowedAttributes from './private/notificationsAllowedAttributes';
 
 import authentication from 'utils/authentication';
-import { UNAUTHORISED_WRITE } from 'helpers/constants';
+import { UNAUTHORISED_WRITE } from 'utils/errorConstants';
 import reqSecure from 'utils/reqSecure';
 
 export default function remove(req) {
