+++ server/api/actions/payments/delete.js
@@ -1,5 +1,5 @@
 import authentication from 'utils/authentication';
-import { UNAUTHORISED_WRITE } from 'helpers/constants';
+import { UNAUTHORISED_WRITE } from 'utils/errorConstants';
 
 import { datumRemove } from '../datum';
