+++ server/api/actions/sessions/delete.js
@@ -1,7 +1,7 @@
 import { datumRemove } from '../datum';
 
 import authentication from 'utils/authentication';
-import { UNAUTHORISED_WRITE } from 'helpers/constants';
+import { UNAUTHORISED_WRITE } from 'utils/errorConstants';
 import reqSecure from 'utils/reqSecure';
 
 export default function remove(req) {
