+++ server/api/actions/support/delete.js
@@ -1,7 +1,7 @@
-import authentication from 'utils/authentication';
-import { UNAUTHORISED_WRITE } from 'helpers/constants';
-
 import { datumRemove } from '../datum';
+
+import authentication from 'utils/authentication';
+import { UNAUTHORISED_WRITE } from 'utils/errorConstants';
 
 export default function remove(req) {
   return new Promise((resolve, reject) => {
