+++ server/api/actions/users/load.js
@@ -3,7 +3,7 @@ import { datumLoad } from '../datum';
 import usersAllowedAttributes from './private/usersAllowedAttributes';
 
 import authentication from 'utils/authentication';
-import { UNAUTHORISED_READ } from 'helpers/constants';
+import { UNAUTHORISED_READ } from 'utils/errorConstants';
 import reqSecure from 'utils/reqSecure';
 
 export default function load(req) {
